+++ CHANGELOG.md
@@ -947,3 +947,4 @@ This is a fast forward to v1.6.3 of daviscook477's fork with a few additional ch
 * Remove unnecessary log in memory churn patch (kiooeht)
 
 #### dev ####
+* Fix crash when running multiple instances of the game (kiooeht)